Harness: squash test cases
