+++ composer.json
@@ -66,7 +66,7 @@
     },
     "autoload-dev": {
         "psr-4": {
-            "Temporal\\Tests\\Acceptance\\": "tests/Acceptance",
+            "Temporal\\Tests\\Acceptance\\App\\": "tests/Acceptance/App",
             "Temporal\\Tests\\Interceptor\\": "tests/Fixtures/src/Interceptor",
             "Temporal\\Tests\\Workflow\\": "tests/Fixtures/src/Workflow",
             "Temporal\\Tests\\Activity\\": "tests/Fixtures/src/Activity",
