--- tests/Acceptance/Harness/Activity/Basic/FeatureActivity.php
@@ -1,18 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Temporal\Tests\Acceptance\Harness\Activity\Basic;
-
-use Temporal\Activity\ActivityInterface;
-use Temporal\Activity\ActivityMethod;
-
-#[ActivityInterface]
-class FeatureActivity
-{
-    #[ActivityMethod('echo')]
-    public function echo(): string
-    {
-        return 'echo';
-    }
-}
